Big rewrite, adding more grid functionality.
The basic idea is that the game provides cartesian coordinates
(+x is East, +y is Up, +z is South from 0 0 0),
but the WorldBorder plugin considers x/z-distance from spawn
(not necessarily at 0 0), which makes a polar coordinate system
more informative.

This patch adds an option to display either the in-game x/z grid
(at a step size of 500m) or the polar grid as measured from spawn
(in 500m and 45° steps).